+++ examples/secrets-cross-account/variables.tf
@@ -1,11 +1,17 @@
-variable "app_account_id" {
-  description = "AWS account ID of the App W (consumer) account that may assume the reader role."
+variable "prod_account_id" {
+  description = "AWS account ID of App W's prod account (may assume the appw/prod reader role)."
   type        = string
   default     = "222222222222"
 }
 
+variable "staging_account_id" {
+  description = "AWS account ID of App W's staging account (may assume the appw/staging reader role)."
+  type        = string
+  default     = "333333333333"
+}
+
 variable "app_workload_role_name" {
-  description = "Name of App W's workload role (Lambda/EC2/ECS task) that needs to read secrets. Lives in the app account."
+  description = "Name of App W's prod workload role (Lambda/EC2/ECS task) that needs to read secrets. Lives in the prod app account."
   type        = string
   default     = "appw-workload"
 }